requirements.txt
pygame>=2.5.0
PyOpenGL>=3.1.6
numpy>=1.24.0
pytest>=7.0.0
pytest-cov>=4.0.0 